First round of updates for Silicon Valley 2014
Some copy changes have been made. I fixed a typo in the global header
that caused the link to "about" to drop a user out of the current Web
Perf Days event that they had chosen to read about.

Google Maps embed was added to location. Location was updated with the
address to Google campus.

EventBright widget was added to the 'register' page. Schedule was
updated to reflect current speakers. Sponsor images were added to both
the index and sponsor pages, respectively.

An updated Google logo and a Fastly logo were added to the images
directory.